server/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary>=2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.3
python-multipart==0.0.20
python-dotenv==1.0.1